Implement watching of related objects
When a related object is changed, the parents that signalled interest
in that object earlier should be notified of the change

Currently this calls the customize for all parents every time a
potentially related object changes. In the future this should be
avoided and only the cached response of the customize hook should be
considered, because we have to find the parents interested in a
related object before enqueueing the parent and calling the customize
hook slows down the main loop, while it should actually only be done
in a worker. Only using the cached response shouldn't be a problem,
because a well behaved parent shouldn't change its interests in
related objects just because a related object changed that it wasn't
interested in in the first place.