Cache responses to the customize hook
The result of the customize hook is expected to only change if the
parent object's spec changes. We use this and cache the response to the
customize hook together with the generation of the parent object
it was based on. We can then use the cached response as long as the
generation of the parent doesn't change.