Notify parent when a related object is becomes uninteresting
The previous implementation would re-sync a parent when a related
object changed to a state that is interesting to the parent. The new
implementation notifies the parent when the updated object was
interesting before the update or is interesting after the update (or
both). This covers the case where an object was interesting to a
parent, but is updates to not be interesting to the parent anymore. In
that case the parent as synced as if the updated object has been
deleted.